Zakonczona czesc pierwsza: rejestracja, logowanie, wylogowanie, dodawanie uzytkownika i adresatow

+++ UzytkownikMenedzer.cpp
@@ -38,7 +38,7 @@ int UzytkownikMenedzer::pobierzIdNowegoUzytkownika() {
 
 bool UzytkownikMenedzer::czyIstniejeLogin(string login) {
 
-    for (int i = 0; i < uzytkownicy.size(); i++) {
+    for (size_t i = 0; i < uzytkownicy.size(); i++) {
         if (uzytkownicy[i].pobierzLogin() == login) {
             cout << endl << "Istnieje uzytkownik o takim loginie." << endl;
             return true;
@@ -48,7 +48,7 @@ bool UzytkownikMenedzer::czyIstniejeLogin(string login) {
 }
 
 void UzytkownikMenedzer::wypiszWszystkichUzytkownikow() {
-    for (int i = 0; i < uzytkownicy.size(); i++) {
+    for (size_t i = 0; i < uzytkownicy.size(); i++) {
         cout << uzytkownicy[i].pobierzId() << endl;
         cout << uzytkownicy[i].pobierzLogin() << endl;
         cout << uzytkownicy[i].pobierzHaslo() << endl;
